更新了 EGE 库。 修复了 InputBox 缩放时文本位置异常的 BUG。 修改了 widget 的测试。

+++ include/Widget.h
@@ -1132,8 +1132,8 @@ private:
 
     std::wstring contentW;
     std::vector<std::wstring> lines;
-    int textWidth = 0;
-    int textHeight = 0;
+    double textWidth = 0;
+    double textHeight = 0;
     TextAlign align = TextAlign::Left;
     int lineSpacing = 0; // 额外行距
     double lastFontScale = -1;  ///< 缓存上次字体缩放比例
